Set balance for inf/NaN to 10; clear unused text
For the balance thing -- this is so it's easier to tell when
a sample is messed up. i'll fix the whole inf/nan thing soon so
this won't matter within a few days.

The unused text being cleared is the sample fancy equation thing.
The reason for that is so that it's clearer that the change was
caused by the multi-assoc stuff (although I guess it'd be worth
clearing the text boxes for multi-assoc after doing a single-assoc
too; something to think about i guess).