Add select microbes file upload + zero-fill value
Lot of changes. as far as I can tell this can replicate every
scatterplot in the notebook (a few use an 0.65 value for
zero-abundance taxa, hence adding an option to specify that) with
the exception of the malassezia/propionibacterium one. gotta test
that out tomorrow -- maybe the notebook's just old, or i'm
overlooking an obvious error. anyway aside from that this is good
progress