Update rank plot colors on single-microbe select
Not the most efficient, but I'm not sure if there's a more direct
way to do this. also it works so ain't gonna prematurely optimize
this, at least not now

granted: still gotta make selection work in the context of an
interactive rank plot (probs through the click once to set numerator,
then click again to set denominator thing for now at least).
if I can't make that work or it looks bad then i'll just make the
rank plot nonzoomable for the time being.