Filter out unplottable samples + restore later
For some reason, I seem to need to make each operation (insert()
then modify() then remove()) run separately in order for this
to work. Chaining them within change() makes the filtering seem
to fail for some reason.

I'm not... especially happy with that workaround, so I think I might
tackle doing this by precomputing the balance changes and then
running:
        remove() -- remove bad samples
        insert() -- insert previously-bad samples from last plot
        run()
        modify() -- update balances
        run()
I guess the advantage this has is it reduces the amount of incomplete
plots that the user has a chance of seeing for a fraction of a
second. However... perhaps another way around this is possible,
through making each regeneration of the sample plot regenerate
it entirely via altering the JSON. That would definitely work, but
it sounds like kind of a hassle. So it's something to put on the TODO
list I guess.