Auto-set sample balances to NaN
Because Vega-Lite filters out null and NaN values, according to
the docs for invalidValues here:
https://altair-viz.github.io/user_guide/configuration.html

I actually uh didn't know that until just now. I might be able to
just set the invalid (inf/-inf/etc) values in the JS to null or NaN
to avoid the multiple run thing (#4).