Replace imminent invalid balance with NaN: fix #4
1) We check the numerator and denominator for if they're <= 0 before
taking the log of them. One thing worth noting: we apparently skated
past a possible error where, in JS, n / Infinity (where n is a normal
number) is equal to 0 (or -0, depending on sign stuff, but -0 === 0
in JS so that part isn't super important). This would have led to
erroneous (and inconspicuous) 0 balances in the sample plot.
However, since we take the difference of the log of the numerator
and the log of the denominator, we get something like n - Infinity,
which evaluates to Inf or -Inf so that'd be manually filtered. (Also
thinking about this more, we never would have been doing
log(a)/log(b) for log ratios -- it would've been log(a/b) or
log(a) - log(b) -- so this is less of an error than I'd thought. Still
something worth writing down in this wall of text for future reference
tho.

2) If either numerator or denominator is <= 0, set NaN as the balance.
As explained in comments, Vega-Lite's default behavior filters out NaN
and null but not Infinities (idk why). I wasn't aware of this until
now, but making use of it is a convenient way to deal with deal with
the problems documented in #4.

3) Reduced the amount of redundant code between updateBalanceSingle()
and updateBalanceMulti() (mostly by creating computeBalance(),
which does the work of NaN-replacing and actually computing balance).

4) Changed logSumAbundancesForSampleTaxa() to just sumAbundances...().
Deferring taking the log lets us do the pre-log <= 0 filtering in
computeBalance().

5) Removed a lot of the old infrastructure for manual filtering. This
reduced a lot of clutter in the file, so I'm cool with that.

(...my bad for kinda breaking the whole "commits should be atomic
changes" principle here)