Add exact rank search support
Replaces the options to search at beginning/end w/ textual queries,
since those aren't especially useful IMO. (If they get requested in
the future, it'll be easy to add them back -- code's in the repo
history.)

This fixes #12 -- we could maybe add fuzzy searching in the future,
but that's a separate issue.

Also fixes #8, because we removed the beginning/end searching.