Actually activate conda env in .travis.yml
...this was a dumb error

Also modified the environment name to be easier to type (just "rrv").
Added related documentation to the Makefile.

(Also explicitly labelled the test and run targets as .PHONY in the
Makefile, a change I figure isn't important enough to merit
purposefully delaying it just for the sake of commit atomicity.)