Remove NumPy as a (direct) dependency
Since it turns out all we were directly using it for was just
creating an arange of integers (which we could just do with a normal
python range).

Also added copyright info to _rank_plot (plus a few minor style
tweaks in comments, etc) -- I guess I forgot to add that until now.