Fix Q2 Metadata obj bug
This was introduced by the earlier change to process_input()
that removed metadata as an argument/return value entirely. It's
true that process_input() didn't actually do anything to the metadata
except return it, but when the Q2 _method code called process_input() it
called it with sample_metadata.to_dataframe() as an argument -- so when
we omitted the process_input() call from the Q2 code, we also omitted
this transformation to a dataframe, which is why we were getting an
error from QIIME 2 about how the Metadata type isn't subscriptable or
whatever (because we were treating it like a dataframe afterwards).

The standalone script code doesn't suffer from this problem because
that immediately calls pd.read_table() on the input sample metadata.