Revert "Revert "Make cmdline arg "_" vs. "-" consistent""
This reverts commit 95e8293196925cde70dec132e3e5f425fbbaa7e7.

Which in turn reverts an earlier commit. The reason I was getting
"arguments not found" or whatever errors from QIIME 2 was because
I tried changing the Q2 arguments, failed, and forgot to reinstall
rankratioviz on my system. Having the standalone use hyphens is a
good idea, since that's what Q2 ends up presenting the user with.