Unify support files loc, and make code work w/ it
Pretty happy with how the code turned out.

This involved removing the extra nested directories within the
output directory, as well as the extra index.html file for some
reason that kept getting created within the qzv files (because
it was set up so that there was, like, an index.html file in
the data/ folder of the qzv archive and then another index.html
file in the rank_plot or support_files subdirectory of data/).
I thought at first that having the index be above everything else
was necessary for Q2View somehow, but apparently not!

uhhhh another thing worth noting: the codebases for _method and _plot
came pretty close together, since now they're doing really similar
stuff. It is probably worth just abstracting this functionality to
another class/function/whatever, so that the Q2 and standalone
versions can just call that. (the only major difference that would
prevent us from literally just yanking out some of this code right
now is that the index path needs to be saved specially or something
in Q2? But we could easily expose that as an option to whatever
abstraction we set up.)

Also, I think I just noticed a bug in searching in RRV -- gonna
check it out. Kinda done for the night but I'll record it at least.

One last thing: i've had the "cory in the house" theme song stuck
in my head for like the past three hours while working on this
and i think i've actually gone crazy at this point