Temporary fix for #7
Eventually we'd want to store all the metadata we have in the JSON,
and when that happens we'll want to store how much metadata we have
in the JSON itself. But for now, just one column (the selected metadata
category), plus "index" and "balance", is being added to the dataset,
so it's sufficient to just start at "3".