Fix #45 and fix #40
JS issues: the metadata column start point thing was the reason
that #45 was broken (the search excluded column 3, which was actually
a taxon), and the | thing was the reason that the most recent
complaint discussed on #40 was broken (see that issue for context).

I still need to fix the empty search thing, as discussed on #40 -- but
that's actually already talked about in #3. I'll try to get to that
soon.