Abstract shared code btwn standalone & Q2 scripts
Closes #37.

Really happy with this solution -- the code is so much more
manageable now!

Worth noting that I had to change one thing to get this to work
-- when determining support_files_loc, I had to remove the ".."
between loc_ and 'support_files'. This is due to the abstracted code's
usage of __file__, which lets us get the current directory of a
program: since generate.py is up a directory from q2/_method.py and
scripts/_plot.py, we have to use a slightly different path to access
support_files/.

I added a comment describing this to the abstracted code, for
posterity.

We could *probably* abstract a bit more stuff (e.g. the call to
process_input()), but that's small fries compared to this change.