Fix test output removal line in Makefile (again)
Now it uses rm -rf instead of just rm -r. See code for a link
to a link explaining for why this is useful.

tldr: this was complaining when the test output directory was
already empty (so nothing was there to be removed), and -f shuts
that up.

And the reason we moved the rm line to above the line that runs the
tests is that, when make encounters an error within a target, it just
quits -- so this meant that tests that errored out wouldn't have the
test output directory deleted, which would lead to more tests failing
but now because the output directory locations already existed.......

yeah programming can be a hassle sometimes :P