Generalize rank plot JSON validation; work on #51
For #51, I added in two functions (ordination_to_rank_dict()
and beta_csv_to_rank_dict()) to beta.csv that each convert songbird
and DEICODE output, respectively, to a standard "format." (note that
the ordination function will need to be tweaked a bit to accept
an existing skbio.OrdinationResults object if it's being called
from the Q2 DEICODE use case, but that's super minor.)

These functions (well, at least the ordination_... one) are already
in use in our testing codebase. However, they should be usable
in the main code of rankratioviz: and, with some tweaking to use
their output ranked_features dict as the reference for all feature
ranking information, this will make it very easy to add support
for songbird input alongside DEICODE input. (Furthermore, it'll make
it super easy to support other file types down the road -- just define
a your_filetype_here_to_rank_dict() function, and rankratioviz can use
that.)

*most* of the byrd test should be working, also. rewrote some of that,
hence the changes/simplification in that file. the reason it's still
commented out is because rankratioviz doesn't, well, actually support
beta.csv files as input. yet. (But once it's ready, we should be able
to just uncomment this test and it should work properly :))

OK!!! I think that's everything! Now to push this to travis and watch
everything fail spectacularly jk hopefully not