Separate sample metadata and feature cts: fix #39
This should get things ready for #31 nicely.

The code now has a good comment explaining it, but the tldr is
that now we only associate metadata with the sample dataset. Feature
counts and their column names are stored in another dataset within
the sample plot JSON (and these dataset names have a prefix that
should prevent them somehow being overwritten -- #55).

Also! The sample plot now is just referred to as sample_plot or
sample_chart, instead of sample_logratio_plot or
sample_logratio_chart. the initial reason I did it the long way at
first was to prevent confusion (e.g. to make clear that "no this isn't
just a sample Vega plot I found somewhere, this is a plot of
samples"), but I think that it's definitely clear by this point.

ALSO! The sample ID is now referred to as just "Sample ID" (since
we're only applying column mapping to the feature columns). This means
that #23 should be a lot closer: now the tooltip will say something
like "Sample ID: 333" or something instead of just "0: 333". I'd still
like to make the headers in the rank plot clearer, but maybe that's
for another day.

In order to restore the prior available functionality (since the web
interface is going to be broken right now due to not knowing where
the counts are), I need to alter the web interface to:
  -Properly relate features from the rank plot to the sample plot's
   JSON's data
  -use above to fix feature searching
  -use above to fix select microbes (features) stuff

And in order to complete #31, I need to alter the web visualization
interface so that it:
  -Can detect all of the metadata columns (just look at the entries
   of a sample point)
  -Has controls for changing the x-axis metadata column
  -Has controls for changing the color metadata column