Use shorter+easier var name for feature col names
Instead of ssmv.samplePlotJSON["datasets"][ssmv.col_names]...
which was being recomputed like twice per search or something...
we just use ssmv.feature_col_names, and it's only set once (in
ssmv.makeSamplePlot()).

Granted: if we need to recreate the sample plot multiple times to
get #31 working, we might need to reconsider this. (but we'll cross
that bridge when we get to it.)