Add metadata selection for x-axis+color of s. plot
Really close to being done with #31!!!

A few more TODOs I want to hammer out before I'm satisfied:

-adding in the option to change the "type" of these encodings. We
 could hook this up to another signal and let the user handle it,
 or there might be d3/vega-esque ways to go about autodetecting that
 kind of stuff.

-avoid setting titles for things that will be overwritten -- want
 to check this

-various testing just to quickly confirm that this does in fact do
 what we want it to (I'm pretty sure it does, but can't hurt to spend
 more time looking into it)

To close the book on #31's discussion from earlier of ways to go about
this: we were able to use the "patch" parameter of vega-embed to
edit the Vega spec by itself, which was a lifesaver. That, in
conjunction with some examples of doing this sorta thing (and really
of just using signals in practice) -- see the added code for links --
made this possible.

tldr: YEET