Abstract h5py dependency to biom-format
I wasn't specifying the "extra" functionality needed properly when
I tried this two weeks ago: you need to reference the actual "extra"
thing needed (in this case, 'hdf5'). Before, I was just listing
biom-format as a requirement like 'biom-format[extra]', which sounds
ridiculous in retrospect but apparently that seemed like a reasonable
thing to me two weeks ago.

ANYWAY this is good because it's one less version thing we have to
keep track of