Close #52 +rename accordingly; incl example output
The example-output-including is done by just adding a "remove
everything in the output directory" thing at the start of the
example notebook. This makes it so you can actually run the thing
multiple times.

By "rename accordingly," I mean that instead of just saying "DEICODE
example" we now actually refer to it as a sleep apnea study.