Fix an issue with rank sorting (#46)
Something silly -- since I was just fetching the rank values directly
from the JSON (without Vega in the middle), these values were getting
treated as strings (so the comparison function I defined when sorting
them didn't work). This sort of thing has happened before, so fixing
this was a lot easier than then.

This emphasizes the importance of following through with #62. The less
there is to worry about, the better.