ENH: Embed JSONs in JS without using CORS: fix #72
Since the JSON files are just loaded within a "script," we don't
need to fire any XMLHTTPRequests or anything like that. This means
rankratioviz' visualizations can be viewed without the need for
a server (i.e. via qiime tools view)! This makes both standalone
and Q2 output a lot easier to inspect.

:D:D:D

NOTE HOWEVER that this will totally break all of the tests, since
they rely on "rank_plot.json" and "sample_plot.json" files being
available in the output directory. We can address this by altering
the tests to extract the JSON from the JS, which is autogenerated
anyway (so extracting it should be easy).