TST: Make tests work with new JSON output
i.e. looking at the actual json object directly instead of at
the file (and then loading the json from there). This actually makes
the code cleaner, which is cool.

this made me realize that a lot of the code in these tests is super
abstractable. should be v easy to abstract more stuff to
testing_utilities, and then make plugging in datasets easy