TST: Use == instead of "in" for entire error msgs
For the "all samples unsupported" case, where we know what the
entire error message is

(For the unsupported feature(s) case, we don't check that the error
message prefix worked out because we don't have an easy(ish) way
to pass the number of features (included in the message prefix)
to testing_utilities. We could take time to do that, but the number of
features is easily obtainable from the dataframe's .shape property
so it's not like that should fail on us. So it's fine.)