MAINT: Initial progress towards RequireJS refactor
Code is a lot more modular ("ssmv" has been done away with and
replaced with a RRVDisplay class, which makes the code a lot easier
to look at and test), and is even split up among multiple JS files.

This doesn't work quite yet (currently changing the plots just
vanishes all of the rank bars for some reason, lol) -- I'll work on
that tomorrow. Since things are so much more easily testable now,
creating tests while I work on restoring the old functionality will
be a lot easier.

PHEW

I'm excited! This is going to be really cool.