MAINT: Whoops remove main.js file
This gets autogenerated by generate.py now (it's the renamed version
of plots.js from before). It was committed due to me copying stuff
in from the environment I was working in.

(If you want to check it out, this was created from the DEICODE
sleep apnea data currently used in some rrv tests and examples! It's
pretty big, though, so it really shouldn't be in this repo.)