TST: Get JS test setup working w/ refactor #82, #2
NICE

Also added the currently-used-by-rrv's-code DOM elements to the test
index.html page, for help with future tests.

I guess *ideally* we wouldn't pollute the global namespace with
mocha's describe/it/etc. stuff, and *ideally* we'd import the Vega*
stack as modules via require, but for now this is totally fine.