MAINT: check searchType; decl a var earlier
jshint was giving an error about rankArray's scope (although it was
never going to be used unless searchType was "rank"), since there
are multiple conditionals in filterFeatures() that check searchType.

In any case, this should be addressed now. ALSO: we check to make
sure that searchType is one of "rank" or "text", so that future
devs can't do something like pass in "asdf" as searchType and
encounter inconsistent behavior (due to if statements that check
"rank" vs. "text" first inconsistently).