MAINT: Update Vega* stack + include via RequireJS
Rather than through individual <script> tags.

Required a bit of fussing to get this to work (a few shims, etc.),
but overall fairly happy with this.

NOTE that the Vega* libs are now at their latest (ish) versions
(Vega 5.3.5, Vega-Lite 3.2.1, Vega-Embed 4.0.0) as of today (April 21,
2019). This produces a warning in the dev console about how the
Altair Vega-Lite spec is older than the Vega-Lite version used
(since Altair only outputs to Vega-Lite version 2.6.0 as of writing),
but 1) this doesn't seem to cause any problems and 2) I imagine this
will be fixed with future Altair releases.

(The reason we updated the Vega* libs here was that this was needed
to make them work nicely with RequireJS. See
https://github.com/vega/vega-embed/issues/8 for context.)