MAINT/TST: Validate sample IDs; add tests re: that
Also added a test of one feature / another in the updateBalanceMulti
tests.

The validation stuff will slow down the UI a bit. We can remove it if
it's needed, but since it's just basically an extra if statement or 2
on top of every updateBalance call... I think it's ok. Shouldn't be
a devastating amount of overhead.