BUG: For now, use balance as default s.p. x-axis
SO: from what I can tell, we *need* to have a quantitative field
be the "initial" x-axis field in the sample plot. Otherwise, updating
the sample plot view just straight-up won't work for some reason.
(And the default_metadata_col variable isn't guaranteed to be
quantitative.)

A temporary solution is using the balance as the default x-axis field.
This field is guaranteed to be quantitative (since we're generating
it) and it's guaranteed to be available.

Should probably update README accordingly?