ENH: Add 2em of margin between the rnk & smp plots
If nothing else, this makes the "actions" dropdown for the rank plot
not get covered by the sample plot stuff as it was before.

Of course, this doesn't resolve any of the truncation issues observed
in #44. Because that'd be too easy ;)

uh also while I'm at it, this commit includes a comment in the CSS
with some code that I use a lot for debugging weird CSS stuff (tm).