ENH: Make TSV output handle spaces + "s correctly
Should fix the problem mentioned in #87. Checks all sample IDs and
(if applicable) the current metadata field values for spaces or quotes
and, if there, encloses the string in question in double-quotes and
replaces all " characters in the string with "".

This should make the TSV files produced by rankratioviz usable in
most environments, since they should generally conform to the
excel-tab TSV "dialect" used by the python csv module (and, in
turn, by the QIIME 2 metadata code).

(For #87 to be done, I still want to add in a method of exporting the
currently selected feature information. Maybe as another TSV file?
IDK.)