ENH/DEV: Integrate debug logging to rankratioviz
This should make it a lot easier to figure out what's going wrong.

And unlike calls to print(), we can leave these in the code.

Added a -v (aka --verbose) flag to the standalone script that
triggers these messages. They're enabled 100% of the time when
running rankratioviz via QIIME 2, which is actually great since that
means they'll only be shown to the user if there's a problem (they'll
end up in the Q2 output logs) or if the user specifies --verbose when
using q2-rankratioviz (they'll be written to stdout, like what happens
when the user passes --verbose to the standalone script).

This worked surprisingly well!