TST: Update JS test V-L specs re: Altair v3 bump
that this needed to be done underscores the importance of following
through with #89. At this point, all that's needed to be done is
just copying in the matching test output to various places. I could
probably write a script to just do that after the python tests are
run, and then call that from the Makefile's pytest directive.

maybe just search test JS files for "var samplePlotJSON = " or
"var rankPlotJSON = " and update those? Would probs need to add
some comments explaining "pls don't change this" lol