ENH: Add tentative boxplot support! #11
This is still kinda buggy -- I'm sure there some strange corner
cases I'm missing that will break this. Worth testing in depth over
the coming days. But still cool that this is feasible!

V-L's need for boxplots to not have unique color/tooltip properties
kinda makes this hard to implement, but it's doable.