ENH: Set labelBound=True for rank plot axes
This should stop (most) of the annoying jittering that happens when
you zoom around in the rank plot. This is introduced because we have
resize=True set for the rank plot's autosize options -- that's needed
(as far as I can tell) in order for the y-axis label not to get cut
off when the y-axis tick markers get long (e.g. due to zooming in
super far, when multiple digits of precision are needed to make the
position clear).

However, a side effect of resize=True is that it takes effect when
we're just moving stuff around -- Vega* is terrified that us moving
an axis label up a bit will cut off the title, so it shifts the title
up, which just looks clunky.

labelBound should fix most of this, but there's still a tiny threshold
where the title gets shifted up. It's definitely better now,
but it'd be a good idea to look into alternative methods for
addressing this in the future.