MAINT: Change r.plot w r.field; refactor destroy()
:D

This basically addresses #46. All that's left is removing the explicit
rankratioviz_x values from the python generate code, and I guess
figuring out why the rank plot is so far on the left now lmao.

Also should test that the rank plot JSON is being adjusted accordingly,
but I'm content to roll that under #2 -- this is basically making Vega*
do all the work for us, which is great.