MAINT: rm explicit rankratioviz_x+simplify rp code
Now, no sorting is done based on the first rank -- Vega-Lite does all
the work for us. This saves some time, and it has the added benefit of
making the tests a bit simpler (fixing them kinda sucked tho lol).

Also, I was able to delete a lot of code by just vectorizing the
addition of the "Classification" column to rank_data.

A few other tiny changes, but IMO the only super important one left to
mention is that we explicitly create a deep copy of V in
gen_rank_plot() now (#63).

At this point, #46 is almost done. All that's left is figuring out why
the rank plot is getting pushed over to the left of the display lol