ENH: quantitative -> ordinal scale for x-axis #1
This isn't as inconvenient as when I tried it before, due to #110
having been resolved (so now the excess horizontal parts of the plot
are scrollable).

However: it's still sorta inconvenient, esp. compared to what we had
before (i.e. an overview of the entire bar plot, albeit a necessarily
sorta inaccurate one due to the use of subpixel sizes).

What I think we should have before we're done with #1 is a
checkbox/toggle/some sort of control that lets the user switch the
rangeStep of the rank plot. We only really need two values:

- Just 1 (i.e. what we have now)
- The value of (chart width in px) / (# of ranked features) -- aka the
  pixel width that perfectly fits the feature "bars" to the plot.

Maybe a good approach would be to detect if the second value is < 1, and
if so warn the user in the visualization interface.

Alternately, we could just make this be a switch between quantitative
and ordinal displays? Although for whatever reason the ordinal display
with the second value seems more "faithful," at least with the Byrd
Intercept ranks (i.e. it actually shows all the bacteria).

In any case: we just need to give the user control of this sort of
stuff, so they can pick what representation they prefer. (And for
creating publication figures or whatevs, they can use the full-size
version.)

Miiight also be a cool idea to let the user expand the rank plot to fill
the screen in a modal-dialog-esque thing -- this might let us have the
best of both worlds. Sorta.