ENH: Increase max tooltip size in rank plot #75
Not quite done yet. Spacing still kinda messes things up, esp.
with the 100bp sOTU prefixes of the feature IDs.

Maybe it'd be best to use " | " as a separator for feature
metadata, instead of "|"? That'd result in the overflow sOTU getting
cut off, while the rest of the metadata (taxonomy, confidence, ...)
is still displayed.