ENH: Add toggling btwn 1px/fit bar wdth: close #1
Really happy with this. When the bar width is detected to be less
than 1 px, a warning is displayed (really happy w/ the CSS for that,
it actually looks passable :).

Also modified the CSS/HTML a bit to make a reusable "invisible"
class, as well as adjusting the h3 styling a bit to remove
user-agent-imposed margins.

And with that... I think we can *finally* merge this branch back in.