ENH: only increase width on rank plot div
This is because increasing the width on the sample plot div
shifts over the "actions" vegaEmbed element to the right of the
page.

We increase(d) the width on the rank plot in order to make it take
up enough of its "space" in order to actually make the
"text-align: right;" work properly. This wasn't a problem for the
sample plot, since we want that one to align *left* (and it seems
like there's no need for this fix to make that happen).