MAINT: decrease max tooltip td width by a bit
Useful now that feature metadata fields are stored independently.

Primary challenge with this, IMO, is balancing "showing full taxonomy
that is listed as a feature ID without any spaces btwn. taxonomic
ranks" with "not making super-unnecessarily-long-tooltips just because
the feature ID is a sOTU". I think this is an acceptable-ish
compromise for now, but maybe my opinion will change when I try this
out on other screen resolutions.