ENH: Get text searching working w/ #132 approach
Notes:

1) Ignores numeric feature metadata values. In the future, we should
support detecting these types? or at least convert to string before
doing text searching.
2) No "rank search" equivalent yet, but users can still include
semicolons in the search. Might reimplement this if
featureMetadataField is "Taxon" or "Taxonomy" or something, but
not urgent IMO.
3) would be nice to add multiple queries for searching -- e.g.
Confidence > 0.95 AND Taxon contains text "p__Firmicutes".
4) I'd like to include at least the taxonomy in the feature lists in
the textareas -- should be doable to "redo" the annotation process in
JS