MAINT: Reconcile new searching w/ RRVDisplay!
The main new change is that, when doing feature selections, we don't
just refer to the feature ID any more. Now we use the whole "feature
row" in the dataset: that is, a JS object that contains the feature ID
and any metadata associated with it.

This is really cool, since it lets us do all these crazy searching
things and still associate the metadata with the feature (so we can
annotate the feature IDs to make pretty output for the user -- this is
done in RRVDisplay.featureRowListToText() now).

This involved a lot of messing around to get things right. In
particular, the RRVDisplay tests required a lot of overhauling.

I'd still like to add a bunch of tests before merging this back into
master, just to verify in even more detail that stuff works (esp. with
tricky corner cases like null values from missing feature metadata and
numeric feature metadata fields).

And this is still *technically* sort of a downgrade, since it doesn't
have taxonomic rank searching or multiple-taxon searching.

Also I'd like for this to support the type-based searching sooner
rather than later, hopefully...?