MAINT: Add skeleton of text-search type #132
Shouldn't be *too* difficult to modify filterFeatures() to detect
the search type and then apply that, same as before.

As discussed in #132 and the various commit messages that reference
it, I'd like to eventually support "joint" queries where you can
filter on multiple criterion (e.g. "contains this text ... and
contains these taxonomic ranks ... and has a confidence greater than
..."), but that sounds like a ton of work. For now, just getting
back to the previous functionality in a bug-free state (i.e. with
issue #125 knocked out) will be good enough.

...So future issues to make after #132 and #125 are:
    -support searching by ranges on numerical feature metadata fields
    -support joint queries across multiple feature metadata fields