DOC: Send test outputs to website as demos
Closes #143.

Possible TODO: add links to QZV files? Actually that'd be a decent
amount of work, since we'd have to explicitly save the QZV files
and include links to them. Probably not worth the time.

I guess we don't really need to have the q2_* copies of everything
(except moving pictures, since that one is only tested in q2); it'd
be ok to remove those copies in the future, if space limitations
are an issue (or we want to decrease repo size). But I guess I don't
mind the redundancy right now (plus it's nice to show "hey this
works with either Q2 or non-Q2, and here's proof right here").