MAINT: Log # of NaN balance samples selection #92
This is a solid start.

We'd want to also log the number of samples dropped due to having
a non-numeric value on a quantitative scale (either for color or
x-axis). I *think* we could do that by running some similar code
in updateSamplePlotField()/updateSamplePlotScale(), and maybe adding
some logic to ensure that each sample is just counted once (i.e. both
null and non-numeric-metadata-on-a-quant-scale samples would be
dropped), but I'm not sure. Should be doable tho :)