MAINT: Replace innerHTML usage with textContent
Also added a unit test that this results in problematic characters
being escaped in the context of dom_utils.updateSampleDroppedDiv().

In the future it'd be good to somehow add back the <strong>/etc.
formatting to emphasize some things, but probs safest to do that
explicitly in HTML/CSS rather than dynamically.

Might be worth adding more tests that weird characters don't mess
up other things in the display, e.g. the sample/feature metadata
selectors or ranking selectors.