TST: Add test re: weird chars and populateSelect()
As far as I'm aware, this is the only thing (besides the various
uses of innerHTML I just replaced) in Qurro that needs to directly
deal with handling weird characters.

json.dumps() and Vega* should deal with the rest of related issues,
I think. So I'm gonna close #156 for the time being.