BUG/ENH: Treat all diffs + md. vals as pd objects
This closes #116 (the new metadata tests ensure this is solved).

This also brings us like 90% of the way to being done with #101.
What we should do is also filter out just-whitespace data, I guess?
To be consistent with QIIME 2.

Now, we should add tests that the Vega* filtering reflects things
properly (NaNs/nulls/Infinities are being filtered). Should be
doable to explicitly set this in Altair.

Once we've verified that the filtering matches the behavior, we'll
be done with #92 and we can merge this branch back in to master.

Also I just spent like five hours making pandas play nicely with NaNs
and now my head hurts

This commit changed a lot of stuff. The bulk of it was making the tests
adapt to the new object-values stuff, as well as testing the new
functionality.

Would be worth testing that differentials' indices are parsed properly,
I guess.