ENH/TST: Handle only-whitespace md vals #101
This should take care of #101! The way we handle it should be the same
as how Q2 handles it.

Also added some pretty thorough tests that this works properly.

MIIIGHT be worth adding an explicit test of that (i.e. calling
qiime2.Metadata.load() and then assert_frame_equal()), but that is
a problem for tomorrow me