BUG: replace ',",\ with | in fix_id()
This was technically a bug because there was a chance that the user
could pass in a field of, say, all quotes ("'''"). Previously, fix_id()
would convert this field to "" -- which is obviously undesirable.
This way, the fields should at least remain at the same length.

(ofc this is probably a temporary solution, but what can you do)