MAINT/BUG: Make more funcs async; fix small bug
The bug had to do with comparing "old" single-selection features to the
new ones (and if they were the same, then ideally nothing would be
done). Since the features are now stored as JS objects, you can't really
test for equality on the objects; now, the features' Feature ID props
are compared, which actually works.

(Tests will be added soon that showcase that this works :D)