TST: Add test of basic #92 stats for singlesel tst
This is super exciting. Should be doable to make a new JS test that
loads fancy JSONs with various weird metadata vals and asserts that the
DOM is being updated correctly, although I guess it isn't *super*
high priority since that side of things is pretty decently unit tested
already.