TST: Test a rank-srch corner case + clean tst code
This should bring feature_computation.js up to 100% coverage!

This tests the corner case where the rank array of the input text to
rankFilterFeatures() is empty. (We'd never see this previously b/c we
already filter out inputs with just whitespace in filterFeatures(),
but this is still conceivably possible with e.g. inputs of just
non-whitespace separator characters.)