TST: Add basic test on s. metadata vals in JS #92
This should at least ensure that all the weird stuff in the SST
sample metadata file makes it to JS ok. The one test I have added
is essentially a copy of
testing_utilities.validate_sample_stats_test_sample_plot_json(), but
from the JS side of things.

This is already pretty good, but we can do even better. IMO, the main
thing we need to do is to test that filters created for this test's
RRVDisplay object match the decisions of getInvalidSampleIDs().

Once that's done I guess we're ready to finally merge this stuff back
in?