TST: Add #92 test re filters for cat/quant fields
Turns out the bug with "rrv" (actually with the JSON) state was
because I need to parse/stringify them before passing them. whoops.
(That's the sort of thing that #162 should address in the future.)

So I imagine these filtering tests as being sort of like a table
for a boolean operator:

C/C
C/Q
Q/C
Q/Q

So far we've got C/C (both x-axis and color categorical) and, now, Q/C
(x-axis quantitative and color categorical) tested. Just need to
add the other two soon.

After that we need to verify that getInvalidSampleIDs() matches up
with these filters. Then... we're good!