Revert "TST: Use a "dev env" Q2 version in Travis #154"
This reverts commit 1d77e64a589c4d792bd6641e909c745dd6852b83.

Once QIIME 2 2019.7 is released, I'll revert this revert (going
back to testing with the dev version). However, the thing is -- the
dev version currently available has the Differentials type definition
merged in, which causes a conflict with songbird. So uhhh the best
thing to do is keep testing with 2019.4, then switch over once all
of the infrastructure is updated (and at this point we can say that
Qurro only supports QIIME 2019.7).