MAINT: Restructure code to have _df_utils module
This removes a lot of code from generate.py, which is a good thing.

Also moved relevant tests to a test_df_utils module. Might break
this up in the future, IDK.

Also renamed the matching and empty sample filtering test modules to
label them as integration tests (and moved the non-integration
tests inside these modules to test_df_utils).

In the long run, this should make the code more readable