MAINT: reorder process_input stuff; work w/ BIOMs
Turns out that we need to delay matching until after filtering
out unextreme features in order to avoid computing the intersection
of two super-huge indices (EMP has on the order of 200k observations
and trying to match those up is going to be horrible).

And it seems like _df_utils.remove_empty_samples() was super slow
on even SparseDataFrames. To make things easier -- and because I
know from experience that it worked, and let Qurro handle huge
EMP-scale amounts of data -- I reorganized things so that we
first filter out ranks and empty samples, THEN do matching.

This works, but it breaks a lot of the unit tests (and some of
the integration tests that rely on specific error messages that
are now changed). Need to fix these then double check that everything
works properly.

This is a prereq for #58. I have an alg for that sketched out, also.